excute simulation of file/dir making in composite pattern example

+++ Composite/TestTree/TTFile.java
@@ -2,18 +2,45 @@ package LearnPattern.Composite.TestTree;
 
 import java.io.File;
 
-public class TTFile implements DirectoryEntry{
+public class TTFile implements DirectoryEntry {
 
     private File _file;
     private String _path;
+    private String _name;
 
-    public TTFile(String path){
+    public TTFile(String path, String name) {
+
         _path = path;
+        _name = name;
+        create();
+
     }
-    public void create(){
-        _file = new File(_path);
+
+    public void create() {
+        
+        try {
+
+            System.out.println("making file in" + _path + "  -   " + _name);
+            _file = new File(_path, _name);
+
+            if (_file.createNewFile()) {
+                System.out.println("File created: " + _file.getName());
+            } else {
+                System.out.println("File already exists.");
+            }
+
+        } catch (Exception e) {
+
+            System.out.println("An error occurred.");
+            e.printStackTrace();
+            
+        }
     }
-    public void remove(){
+
+    public void remove() {
+
+        if(_file == null) return;
+
         _file.delete();
     }
 }
